WEB - Implementação do carrinho de compras (GeekShooping.Web)

+++ GeekShooping.Web/Services/ICartService.cs
@@ -7,7 +7,7 @@ namespace GeekShooping.Web.Services
         Task<CartViewModel> FindCartByUserId(string userId, string token);
         Task<CartViewModel> AddItemToCart(CartViewModel cart, string token);
         Task<CartViewModel> UpdateItemToCart(CartViewModel cart, string token);
-        Task<bool> RemoveFromCart(long cartId, string token);
+        Task<bool> RemoveFromCart(int cartId, string token);
         Task<bool> ApplyCoupon(CartViewModel cart, string couponCode, string token);
         Task<bool> RemoveCoupon(string userId, string token);
         Task<bool> ClearCart(string userId, string token);
